+++ drift_dev/lib/src/analysis/resolver/shared/data_class.dart
@@ -14,15 +14,20 @@ String dataClassNameForClassName(String tableName) {
   // they're storing (users, products, ...). We try to find the singular word
   // from the table name.
 
-  // todo we might want to implement some edge cases according to
-  // https://en.wikipedia.org/wiki/English_plurals
-
   if (tableName.endsWith('s')) {
-    return tableName.substring(0, tableName.length - 1);
+    if (tableName.endsWith('ss') ||
+        tableName.endsWith('us') ||
+        tableName.endsWith('sses')) {
+      return tableName;
+    } else if (tableName.endsWith('ies')) {
+      return '${tableName.substring(0, tableName.length - 3)}y';
+    } else {
+      return tableName.substring(0, tableName.length - 1);
+    }
+  } else {
+    // Default behavior if the table name is not a valid plural.
+    return '${tableName}Data';
   }
-
-  // Default behavior if the table name is not a valid plural.
-  return '${tableName}Data';
 }
 
 CustomParentClass? parseCustomParentClass(